Se implementa la funcionalidad de los componentes de Proveedores y Compras con todas las operaciones CRUD, adicionalmente se resuelven bugs menores en el backend

+++ Backend C#/InventarioAPI/InventarioAPI.Infraestructura/Database/Entidades/CompraEntity.cs
@@ -10,7 +10,7 @@ namespace InventarioAPI.Infraestructura.Database.Entidades
         public int id_compra { get; set; }
         public int id_cliente { get; set; }
         public DateTime hora_de_compra { get; set; }
-        public float valor_total { get; set; }
+        public decimal valor_total { get; set; }
 
     }
 }
